Move PopulateSymbolTableWithName into ICSharpBinder
Avoid test-type-and-cast chain.
